feat (Mapping): reworks BlockReferenceComponentMapDefinition
This map definition now uses IComponentMapDefinition instead of ComponentMapDefinition. This removes the ability to directly map to a IPublishedElement. This is carried over from recent changes to the BlockReferenceSubcomponent map definition doing so reduces class inheritance and also focuses the mapping to just use cases where we use BlockGridItem or BlockListItem.

Mapping directly to a IPublishedElement can still be achieved through using the ComponentMapDefinition.